+++ lib/generators/shopify_app/install/install_generator.rb
@@ -11,10 +11,6 @@ module ShopifyApp
       class_option :embedded, type: :string, default: 'true'
       class_option :api_version, type: :string, default: nil
 
-      def add_dotenv_gem
-        gem('dotenv-rails', group: [:test, :development])
-      end
-
       def create_shopify_app_initializer
         @application_name = format_array_argument(options['application_name'])
         @scope = format_array_argument(options['scope'])
@@ -40,12 +36,19 @@ module ShopifyApp
       end
 
       def create_embedded_app_layout
-        if embedded_app?
-          copy_file 'embedded_app.html.erb', 'app/views/layouts/embedded_app.html.erb'
+        return unless embedded_app?
+
+        copy_file 'embedded_app.html.erb', 'app/views/layouts/embedded_app.html.erb'
+        copy_file '_flash_messages.html.erb', 'app/views/layouts/_flash_messages.html.erb'
+
+        if ShopifyApp.use_webpacker?
+          copy_file('shopify_app.js', 'app/javascript/shopify_app/shopify_app.js')
+          copy_file('flash_messages.js', 'app/javascript/shopify_app/flash_messages.js')
+          copy_file('shopify_app_index.js', 'app/javascript/shopify_app/index.js')
+          append_to_file('app/javascript/packs/application.js', 'require("shopify_app")')
+        else
           copy_file('shopify_app.js', 'app/assets/javascripts/shopify_app.js')
-          copy_file '_flash_messages.html.erb', 'app/views/layouts/_flash_messages.html.erb'
-          copy_file('flash_messages.js',
-            'app/assets/javascripts/flash_messages.js')
+          copy_file('flash_messages.js', 'app/assets/javascripts/flash_messages.js')
         end
       end
 
@@ -55,6 +58,14 @@ module ShopifyApp
 
       def mount_engine
         route "mount ShopifyApp::Engine, at: '/'"
+      end
+
+      def insert_hosts_into_development_config
+        inject_into_file(
+          'config/environments/development.rb',
+          "  config.hosts = (config.hosts rescue []) << /\\h+.ngrok.io/\n",
+          after: "Rails.application.configure do\n"
+        )
       end
 
       private